+++ src/utils.h
@@ -3,5 +3,5 @@
 #include <iostream>
 #include <fstream>
 
-Eigen::MatrixXf vecToMat(std::vector<float> input, int nRows);
-void matToCSV(Eigen::MatrixXf mat, std::string filePath);
+Eigen::MatrixXd vecToMat(std::vector<float> input, int nRows);
+void matToCSV(Eigen::MatrixXd mat, std::string filePath);
